No longer creates 2 files instead of 1
