+++ src/extension.ts
@@ -8,6 +8,7 @@ import { DodonaClient } from "./api/client";
 import { Exercise } from "./api/resources/exercise";
 import { getSyntax } from "./commentSyntax";
 import * as path from "path";
+import { getToken } from './checkToken';
 
 // Initialise a Dodona client.
 let dodona = new DodonaClient("https://dodona.ugent.be");
@@ -204,8 +205,9 @@ export function activate(context: vscode.ExtensionContext) {
         const newFile = vscode.Uri.parse("untitled:" + path.join(vscode.workspace.rootPath, `${exercise.name}.${exercise.language.extension}`));
         vscode.workspace.openTextDocument(newFile).then(document => {
             const edit = new vscode.WorkspaceEdit();
+            const commentedUrl = getSyntax(exercise.language.name, exercise.url)
             const boilerplate = exercise.boilerplate != null ? exercise.boilerplate : "";
-            edit.insert(newFile, new vscode.Position(0, 0), `${getSyntax(exercise.language.name, exercise.url)}\n${boilerplate}`);
+            edit.insert(newFile, new vscode.Position(0, 0), `${commentedUrl}\n${boilerplate}`);
             return vscode.workspace.applyEdit(edit).then(success => {
                 if (success) {
                     vscode.window.showTextDocument(document);
@@ -223,31 +225,6 @@ export function activate(context: vscode.ExtensionContext) {
     // Register & create the tree view for the plugin
     vscode.window.registerTreeDataProvider("dodona-exercises", dataProvider);
     vscode.window.createTreeView("dodona-exercises", { treeDataProvider: dataProvider });
-
-    function getToken(platform: string): string | null {
-        // Get the API token from the settings.
-        const config = vscode.workspace.getConfiguration("dodona");
-        const token = config.get<string>(`api.token.${platform}`);
-
-        // Display a warning notification if no Api token has been set
-        if (!token) {
-            const instructionsButton = "Instructions";
-            const settingsButton = "Open Settings";
-            vscode.window.showErrorMessage(`You have not yet configured your ${platform} Api token. To correctly set up your Visual Studio Code, click the Instructions button below.`, instructionsButton, settingsButton)
-                .then(selection => {
-                    // Open the page when the user clicks the button
-                    if (selection === instructionsButton) {
-                        vscode.env.openExternal(vscode.Uri.parse("https://dodona-edu.github.io/en/guides/vs-code-extension/#_3-insert-api-token"));
-                        vscode.env.openExternal(vscode.Uri.parse("https://dodona-edu.github.io/en/guides/vs-code-extension/#_3-insert-api-token"));
-                    } else if (selection === settingsButton) {
-                        vscode.commands.executeCommand("workbench.action.openSettings2");
-                    }
-                });
-            return null;
-        }
-
-        return token;
-    }
 
     // Return the opposite of the input platform (used in error message)
     function oppositePlatform(inputPlatform: string) {
